transforms update - removes htmlmin - adds jamPack
