+++ .vscode/settings.json
@@ -31,10 +31,12 @@
   },
   "cSpell.words": [
     "alpinejs",
+    "cleancache",
     "divriots",
     "flightdeck",
     "Heltzel",
     "Laserwave",
+    "postbuild",
     "thefold"
   ]
 }
